[ES|QL][Lens] Remove allColumns from state (#174500)
## Summary

The basic feature that this PR offers is the removal of allColumns from
the state. It now retrieves them from the cache when they are needed.

This is going to make the SO much lighter and improve a bit the
performance. This must be bwc.
I also think that it makes the state much easier to understand and the
code so it is a very nice technical improvement.

### Checklist

- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios