[Fleet] enabling flaky tests with endpoint (#174679)
## Summary

Closes https://github.com/elastic/kibana/issues/163203
Closes https://github.com/elastic/kibana/issues/167188
Closes https://github.com/elastic/kibana/issues/156941

Enable flaky tests, using endpoint-8.6.1 as a fixture bundled package.
See discussion:
https://github.com/elastic/kibana/issues/167188#issuecomment-1887346563

Closes https://github.com/elastic/kibana/issues/170690
The es error was reproducible with an API request to index a document. I
had a look at the data stream mappings, and experimented with providing
values for the other keyword `time_series_dimension` fields. It seems
the issue is fixed by providing value for one more field.