[ML] Improving pattern analysis refresh behaviour (#174516)
Improves behaviour when an auto-refresh is enabled or the user presses
then time picker's refresh button.
We now keep a record of the previous event rate data to determine if new
data has appeared in the chart. If so, the pattern analysis is run
again.
As selected time period could always contain the same number of docs, we
need to keep track of the time stamp of the last doc in the chart. I've
updated the query created by `getDocumentCountStatsRequest` to also get
the last doc and to return the timestamp in ms.