[Slo] serverless integration tests (#172786)
## 🍒 Summary
This PR adds basic serverless integration tests for SLO and covers 2
scenarios:
- SLO creation 
- SLO deletion 

There is another PR that adds SLO integration tests for stateful and
covers more scenarios:
- create
- delete
- update
- reset
- get/find

Current PR will cover only the create and delete scenarios. I want to
check the flakiness of these tests, before introducing new ones. Another
reason for not covering all scenarios in this PR is because we don't
want to have duplicate effort with @dominiqueclarke's
[PR](https://github.com/elastic/kibana/pull/173236). Once stateful tests
are reviewed and merged, we can come up with a plan on how/if to
continue with serverless tests and more scenarios.

**TODO**: Create a github issue to track the SLO integration tests
effort. Also we need to add privilege & space related test cases

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>