[DOCS] Adds more information how to handle a breaking change in 8.10.0 - SLOs (#174901)
A breaking change was introduced in 8.10.0, which makes SLOs created
before migration not work. They need to be removed and reset-up. The
release note in 8.10.0 did not make this clear. More information has
been added to the release note in this PR including steps to remove the
SLOs that no longer work.

Closes: #173938