[RespOps][Security Solution] Enable SentinelOne feature flags in both Security Solution and Stack connectors (#174648)
## Summary

- Enables the SentinelOne Feature Flags for both Stack Connectors and
Security Solution


## Summary

- Enable the experimental feature flag for SentinelOne. Features
enabled:

```
xpack.securitySolution.enableExperimental:
   - sentinelOneManualHostActionsEnabled

xpack.stack_connectors.enableExperimental:
   - sentinelOneConnectorOn
```

> [!NOTE]
> This change was [initially done in the `v8.12.0`
branch](https://github.com/elastic/kibana/pull/173859) soon after
feature feature and is not being enabled in `main` as we get closer to
GA of `v8.12.0`. This update to `main` will enable the feature for
serverless when the next release of it is deployed.