[ci] Stop pull request build if Checks step fails (#175269)
This adds a new block containing the `Checks` and `Build Kibana
Distribution and Plugins` steps. If either fails the build will be
aborted before tests run.

Running `Checks` in parallel with `Build Kibana Distribution and
Plugins` step should minimize the increase in time for the pipeline to
finish (est +5 minutes vs +15 minutes).