Use controlled version of InlineEdit component (#174850)
## Summary

The [Inline Edit component](https://eui.elastic.co/#/forms/inline-edit)
did not have a controlled version when we started using it, so we had to
implement some nasty workarounds to get the component to re-render.

The Eui Team has since then added props to make the component
controlled, so we don't need this anymore. This PR cleans up the hacks.
As well as some unused props.