[Lens] Fixes mispelled complementary palette (#176136)
## Summary

Closes https://github.com/elastic/kibana/issues/161194

Dashboard with Lens, aggbased and TSVB using the legacy complimentary
palette.

<img width="2104" alt="image"
src="https://github.com/elastic/kibana/assets/17003240/9cc0da68-3f36-401a-a264-a15e71c46cdd">


Correct usage in the list.

<img width="530" alt="image"
src="https://github.com/elastic/kibana/assets/17003240/4f9046f8-77ae-4d30-b12f-92844d77a758">


I don't want to run a migration. All the new visualizations will have
the correct id in the SO and I am fallbacking for the existing. As with
the new color mapping we are going to construct this palette (it won't
be in the list) I think that this is enough to solve the problem.