[Response Ops][Alerting] Removing unnecessary error log when getting the rules setting saved object (#176538)
Resolves https://github.com/elastic/kibana/issues/174534

## Summary

Removing extra error log that was unneeded and did not check for a saved
object not found error before logging. Since the `getSettings` function
is only used internally by the client, I removed the logger altogether
and let the calling function catch and handle the error.

## To Verify
Start up ES and Kibana and create a rule. On the first run, you should
see 2 INFO logs for creating the settings but no error logs.

```
[2024-02-08T12:53:44.327-05:00][INFO ][plugins.alerting] Creating new default flapping rules settings for current space.
[2024-02-08T12:49:41.780-05:00][INFO ][plugins.alerting] Creating new default query delay rules settings for current space.
```