[Logs Explorer] Add logic to customise the JSON rendering logic (#176695)
## Summary

Closes - https://github.com/elastic/kibana/issues/175091

The content virtual column in Logs Explorer would render the whole JSON
body when none of the message fallbacks were present. With this PR
(Iteration 1), we added a custom filter logic. With this logic, we will
remove certain known fields from the JSON so as to make this column only
render Content Related fields.

The original ticket talks about making this list customisable and adding
persistence. This would be something which we would tackle in follow up
iterations.

## Additional Fix
When i introduced Synthtrace to Logs Explorer Functional Test, i broke
the FTR Context type for our tests, due to which PageObjects and
Services were not inferred properly in the IDE. I took the liberty to
fix them as part if this PR.

## Demo -

![JSON
rendering](https://github.com/elastic/kibana/assets/7416358/a57f3d75-1706-49bd-90ff-1ef08ce2abbd)